Add react-native-url-polyfill
Without this polyfill, calling `server.start()` will result in an Error: not implemented message followed by Error: Invariant Violation: Module AppRegistry is not a registered callable module (calling runApplication). A frequent cause of the error is that the application entry file path is incorrect.

This is due to the React Native URL polyfill that throws Not Implemented exceptions for functions that MSW calls.